品牌删除

+++ mingrui-shop-parent/mingrui-shop-service/mingrui-shop-service-xxx/src/main/java/com/baidu/shop/service/impl/BrandServiceImpl.java
@@ -5,9 +5,12 @@ import com.baidu.shop.base.Result;
 import com.baidu.shop.dto.BrandDTO;
 import com.baidu.shop.entity.BrandEntity;
 import com.baidu.shop.entity.CategoryBrandEntity;
+import com.baidu.shop.entity.SpuEntity;
 import com.baidu.shop.mapper.BrandMapper;
 import com.baidu.shop.mapper.CategoryBrandMapper;
+import com.baidu.shop.mapper.SpuMapper;
 import com.baidu.shop.service.BrandService;
+import com.baidu.shop.status.HTTPStatus;
 import com.baidu.shop.utils.BaiduBeanUtil;
 import com.baidu.shop.utils.ObjectUtil;
 import com.baidu.shop.utils.PinyinUtil;
@@ -16,6 +19,7 @@ import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import com.google.gson.JsonObject;
 import jdk.nashorn.internal.ir.FunctionNode;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.RestController;
 import tk.mybatis.mapper.entity.Example;
@@ -40,13 +44,15 @@ public class BrandServiceImpl extends BaseApiService implements BrandService {
     private BrandMapper brandMapper;
     @Resource
     private CategoryBrandMapper categoryBrandMapper;
+    @Resource
+    private SpuMapper spuMapper;
 
     @Override
     public Result<PageInfo<BrandEntity>> getBrandInfo(BrandDTO brandDTO) {
 
         //分页
-        if(ObjectUtil.isNotNull(brandDTO.getPage()) && ObjectUtil.isNotNull(brandDTO.getRows()))
-             PageHelper.startPage(brandDTO.getPage(), brandDTO.getRows());
+        if (ObjectUtil.isNotNull(brandDTO.getPage()) && ObjectUtil.isNotNull(brandDTO.getRows()))
+            PageHelper.startPage(brandDTO.getPage(), brandDTO.getRows());
 
         //排序
         Example example = new Example(BrandEntity.class);
@@ -54,8 +60,8 @@ public class BrandServiceImpl extends BaseApiService implements BrandService {
         if (!StringUtils.isEmpty(brandDTO.getSort())) example.setOrderByClause(brandDTO.getOrderByClause());
 
         Example.Criteria criteria = example.createCriteria();
-        if(ObjectUtil.isNotNull(brandDTO.getId()))
-            criteria.andEqualTo("id",brandDTO.getId());
+        if (ObjectUtil.isNotNull(brandDTO.getId()))
+            criteria.andEqualTo("id", brandDTO.getId());
 
         if (StringUtils.isEmpty(brandDTO.getName()))
             criteria.andLike("name", "%" + brandDTO.getName() + "%");
@@ -67,18 +73,21 @@ public class BrandServiceImpl extends BaseApiService implements BrandService {
         return this.setResultSuccess(pageInfo);
     }
 
+
     @Override
     public Result<JsonObject> saveBrand(BrandDTO brandDTO) {
+
+        brandMapper.insertSelective(BaiduBeanUtil.copyProperties(brandDTO,BrandEntity.class));
+
         BrandEntity brandEntity = BaiduBeanUtil.copyProperties(brandDTO, BrandEntity.class);
+
         brandEntity.setLetter(PinyinUtil.getUpperCase(String.valueOf(brandEntity.getName().charAt(0))
                 , PinyinUtil.TO_FIRST_CHAR_PINYIN).charAt(0));
 
         brandMapper.insertSelective(brandEntity);
-        this.insertCategoryAndBrand(brandDTO, brandEntity);
 
-        //return this.setResultSuccess();
+        if(brandDTO.getCategory().contains(",")){
 
-        if (brandDTO.getCategory().contains(",")) {
             List<CategoryBrandEntity> categoryBrandEntities = Arrays.asList(brandDTO.getCategory().split(","))
                     .stream().map(cid -> {
 
@@ -92,7 +101,7 @@ public class BrandServiceImpl extends BaseApiService implements BrandService {
             //批量新增
             categoryBrandMapper.insertList(categoryBrandEntities);
 
-        } else {
+        }else{
             //新增
             CategoryBrandEntity entity = new CategoryBrandEntity();
 
@@ -101,10 +110,11 @@ public class BrandServiceImpl extends BaseApiService implements BrandService {
 
             categoryBrandMapper.insertSelective(entity);
         }
+
         return this.setResultSuccess();
     }
-
     @Override
+    @Transactional
     public Result<JsonObject> editBrand(BrandDTO brandDTO) {
         BrandEntity brandEntity = BaiduBeanUtil.copyProperties(brandDTO, BrandEntity.class);
         brandEntity.setLetter(PinyinUtil.getUpperCase(String.valueOf(brandEntity.getName().charAt(0))
@@ -117,22 +127,38 @@ public class BrandServiceImpl extends BaseApiService implements BrandService {
         this.deleteCategoryAndBrand(brandEntity.getId());
 
         //新增新的数据
-        this.insertCategoryAndBrand(brandDTO,brandEntity);
+        this.insertCategoryAndBrand(brandDTO, brandEntity);
 
         return this.setResultSuccess();
     }
-
 
 
     @Override
+    @Transactional
     public Result<JsonObject> deleteBrand(Integer id) {
 
-        //删除品牌
+        Example example = new Example(SpuEntity.class);
+        example.createCriteria().andEqualTo("brandId", id);
+        List<SpuEntity> spuEntityList = spuMapper.selectByExample(example);
+        if (spuEntityList.size() > 0) {
+            return this.setResultError(HTTPStatus.OPERATION_ERROR, "当前品牌被商品绑定,不能删除");
+        }
         brandMapper.deleteByPrimaryKey(id);
-        //关系
+
         this.deleteCategoryAndBrand(id);
+
         return this.setResultSuccess();
     }
+
+
+    @Override
+    public Result<List<BrandEntity>> getBrandByCategory(Integer cid) {
+
+        List<BrandEntity> list = brandMapper.getBrandByCategory(cid);
+
+        return this.setResultSuccess(list);
+    }
+
     private void deleteCategoryAndBrand(Integer id) {
         //通过brandID删除中间表的数据
         Example example = new Example(CategoryBrandEntity.class);
